clang-tidy: Enable bugprone-narrowing-conversions
Address the following warning to allow global enforcement:

```
../src/platforms/rainier/nisqually.cpp:172:20: error: narrowing conversion from 'unsigned long' to signed type 'int' is implementation-defined [bugprone-narrowing-conversions,-warnings-as-errors]
    for (int index :
                   ^
```

+++ src/platforms/rainier/nisqually.cpp
@@ -169,7 +169,7 @@ void Nisqually::detectWilliwakasCards(Notifier& notifier)
     gpiod::chip williwakasPresenceChip(sysfsChip.getName().string(),
                                        gpiod::chip::OPEN_BY_NAME);
 
-    for (int index :
+    for (unsigned long index :
          std::views::iota(0UL, Nisqually::williwakasPresenceMap.size()))
     {
         int offset = Nisqually::williwakasPresenceMap.at(index);
